requirements.txt
altgraph==0.17
certifi==2021.5.30
charset-normalizer==2.0.4
future==0.18.2
idna==3.2
importlib-metadata==4.6.4
pefile==2021.5.24
Pillow==8.3.1
psycopg2==2.9.1
pycryptodome==3.10.1
pyinstaller==4.5
pyinstaller-hooks-contrib==2021.3
pywin32==301
pywin32-ctypes==0.2.0
requests==2.26.0
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0